+++ src/Strategy/Scala.hs
@@ -24,7 +24,7 @@ import Effect.Exec
 import Effect.ReadFS
 import GHC.Generics (Generic)
 import Path
-import Prettyprinter (viaShow)
+import Prettyprinter (viaShow, vsep)
 import Strategy.Maven.Pom qualified as Pom
 import Strategy.Maven.Pom.Closure (MavenProjectClosure, buildProjectClosures)
 import Strategy.Maven.Pom.Closure qualified as PomClosure
@@ -89,10 +89,8 @@ findProjects = walk' $ \dir _ files -> do
 newtype FailedToListProjects = FailedToListProjects (Path Abs Dir)
   deriving (Eq, Ord, Show)
 
--- TODO(warnings): this warning is not helpful
 instance ToDiagnostic FailedToListProjects where
-  renderDiagnostic (FailedToListProjects dir) =
-    "Found an sbt build manifest, but failed to list sbt projects in " <> viaShow dir
+  renderDiagnostic (FailedToListProjects dir) = "Failed to discover and analyze sbt projects, for sbt build manifest at:" <> viaShow dir
 
 makePomCmd :: Command
 makePomCmd =