+++ Changelog.md
@@ -14,6 +14,11 @@
 - Cocoapods: Improves error and warning messages. ([#807](https://github.com/fossas/fossa-cli/pull/807))
 - Golang: Improves error and warning messages. ([#809](https://github.com/fossas/fossa-cli/pull/809))
 - Gradle: Improves error and warning messages. ([#804](https://github.com/fossas/fossa-cli/pull/804))
+- Scala: Improves error and warning messages.
+- Clojure: Improves error and warning messages.
+- Nim: Improves error and warning messages.
+- Rust: Improves error and warning messages.
+- UX: Improves error message when executable is not found. 
 
 ## v3.1.0